Return / in AppContext.BaseDirectory for wasm
