+++ src/libraries/System.Private.CoreLib/src/System/AppContext.cs
@@ -36,7 +36,9 @@ namespace System
                 throw new ArgumentNullException(nameof(name));
 
             if (s_dataStore == null)
+            {
                 return null;
+            }
 
             object? data;
             lock (s_dataStore)
@@ -141,15 +143,6 @@ namespace System
             {
                 s_dataStore.Add(new string(pNames[i]), new string(pValues[i]));
             }
-        }
-
-        private static string GetBaseDirectoryCore()
-        {
-            // Fallback path for hosts that do not set APP_CONTEXT_BASE_DIRECTORY explicitly
-            string? directory = Path.GetDirectoryName(Assembly.GetEntryAssembly()?.Location);
-            if (directory != null && !Path.EndsInDirectorySeparator(directory))
-                directory += PathInternal.DirectorySeparatorCharAsString;
-            return directory ?? string.Empty;
         }
 #endif
     }
